breaking out lights into their own class.
